Safari-611.1.10/LayoutTests/printing/numberOfPages.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="resources/paged-media-test-utils.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="sandbox"></div>
<script>
description("Test for internals.numberOfPages()");

function test()
{
    createBlockWithRatioToPageHeight("firstPage", 0.6);
    createBlockWithRatioToPageHeight("secondPage", 0.6);

    numberOfPagesShouldBe(2);
    numberOfPagesShouldBe(1, 1000, 10000);

    document.body.removeChild(document.getElementById("sandbox"));
}

var successfullyParsed = true;
</script>
<script>runPrintingTest(test);</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
